+++ dfx-library/dfxplugin.h
@@ -478,17 +478,17 @@ public:
 
   	// Overrides to define custom properties. Note that calling these directly will not update listeners;
   	// only the DfxGuiEditor versions do that.
-	virtual long dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex, 
+	virtual long dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex, 
 									 size_t& outDataSize, dfx::PropertyFlags& outFlags)
 	{
 		return dfx::kStatus_InvalidProperty;
 	}
-	virtual long dfx_GetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex, 
+	virtual long dfx_GetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex, 
 								 void* outData)
 	{
 		return dfx::kStatus_InvalidProperty;
 	}
-	virtual long dfx_SetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex, 
+	virtual long dfx_SetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex, 
 								 void const* inData, size_t inDataSize)
 	{
 		return dfx::kStatus_InvalidProperty;
