+++ geometer/geometer.cpp
@@ -188,7 +188,7 @@ void PLUGIN::dfx_PostConstructor()
   getsettings().setAllowNoteEvents(true);
 }
 
-long PLUGIN::dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex,
+long PLUGIN::dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex,
                                  size_t& outDataSize, dfx::PropertyFlags& outFlags)
 {
   switch (inPropertyID)
@@ -206,7 +206,7 @@ long PLUGIN::dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScop
   }
 }
 
-long PLUGIN::dfx_GetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex,
+long PLUGIN::dfx_GetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex,
                              void* outData)
 {
   switch (inPropertyID)
