+++ turntablist/turntablist.h
@@ -142,11 +142,11 @@ public:
 	void processaudio(float const* const* inAudio, float* const* outAudio, unsigned long inNumFrames) override;
 	void processparameters() override;
 
-	long dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex,
+	long dfx_GetPropertyInfo(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex,
 							 size_t& outDataSize, dfx::PropertyFlags& outFlags) override;
-	long dfx_GetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex,
+	long dfx_GetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex,
 						 void* outData) override;
-	long dfx_SetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned long inItemIndex,
+	long dfx_SetProperty(dfx::PropertyID inPropertyID, dfx::Scope inScope, unsigned int inItemIndex,
 						 void const* inData, size_t inDataSize) override;
 
 	OSStatus SaveState(CFPropertyListRef* outData) override;
